client/requirements.txt
paho-mqtt==1.6.1
requests==2.32.3

# You may have to add more libraries
